% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stanPkgTest-package.R
\docType{package}
\name{stanPkgTest-package}
\alias{stanPkgTest-package}
\alias{stanPkgTest}
\title{The 'stanPkgTest' package.}
\description{
A DESCRIPTION OF THE PACKAGE
}
\references{
Stan Development Team (2020). RStan: the R interface to Stan. R package version 2.21.2. https://mc-stan.org
}
